refactor(execution): scope modules into folders
move symbolTable, parser, and interpreter modules into respective folders
